fixed bug reported by @erichoog on PR #61
Adding extra character to the class attribute filled the
buffer. Increase to 1280 fixes it.

@dafky2000 recommends handling arbitrarily sized strings
using pointer reuse. I don't know how to do that without
setting a limit at some point. Having limits seems to help
with detecting really long lines. In this case, it's ok.
There's 10 tags on this, one of the tags is very long. But
at some point, a missing comma or some other unforeseen mishap
might be caught by having a limit set and doing a buf_check.